Move fix to appropriate new place in code
